Cria página de comparação de seus votos com candidatos
