+++ client/src/components/candidatos/CandidatosContainer.js
@@ -100,6 +100,7 @@ class CandidatosContainer extends Component {
       return (
         <Candidato
           key={candidato.cpf}
+          id={candidato.cpf}
           nome={candidato.nome_urna}
           siglaPartido={candidato.sg_partido}
           estado={candidato.uf}
@@ -138,8 +139,8 @@ class CandidatosContainer extends Component {
             <Spinner />
           </div>
         ) : (
-            <div className="candidatos">
-              {/*<div className="row">
+          <div className="candidatos">
+            {/*<div className="row">
               <div className="col-8 col-xs-8 col-md-8 col-lg-8">
                 <input
                   className="barra-filtro-candidato form-control"
@@ -168,9 +169,9 @@ class CandidatosContainer extends Component {
                 </select>
               </div>
         </div>*/}
-              <FlipMove>{candidatos}</FlipMove>
-            </div>
-          )}
+            <FlipMove>{candidatos}</FlipMove>
+          </div>
+        )}
       </div>
     );
   }
